refactor(confirm): merge toggle into confirm with dual-mode support
