+++ confirm/command_test.go
@@ -37,3 +37,39 @@ func TestSetCancelButtonLabel(t *testing.T) {
 		})
 	}
 }
+
+func TestSetCursor(t *testing.T) {
+	tests := []struct {
+		name string
+		cfg  Config
+		want string
+	}{
+		{"default", Config{}, ">"},
+		{"custom", Config{Cursor: "*"}, "*"},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := setCursor(&tt.cfg); got != tt.want {
+				t.Errorf("setCursor() = %q, want %q", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestSetDivider(t *testing.T) {
+	tests := []struct {
+		name string
+		cfg  Config
+		want string
+	}{
+		{"default", Config{}, "/"},
+		{"custom", Config{Divider: "|"}, "|"},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := setDivider(&tt.cfg); got != tt.want {
+				t.Errorf("setDivider() = %q, want %q", got, tt.want)
+			}
+		})
+	}
+}
